Remove `useStripe` hook
Move the useEffect hooks into the common `UsageBasedBillingConfig`
component.
